enhancement: refactoring and modularization
