+++ lib/data/models/notification.dart
@@ -101,10 +101,8 @@ class NotificationModel {
         return NotificationType.message;
       case Action.cantDo:
       case Action.tradePubkey:
-        return NotificationType.system;
       case Action.timeoutReversal:
-        // TODO: Handle this case.
-        throw UnimplementedError();
+        return NotificationType.system;
     }
   }
 }
