+++ lib/features/notifications/notifiers/notifications_notifier.dart
@@ -1,47 +1,18 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mostro_mobile/data/repositories/notifications_history_repository.dart';
 import 'package:mostro_mobile/data/models/notification.dart';
+import 'package:mostro_mobile/features/notifications/notifiers/notification_temporary_state.dart';
+import 'package:mostro_mobile/features/notifications/notifiers/notifications_state.dart';
+import 'package:mostro_mobile/features/notifications/utils/notification_message_mapper.dart';
 import 'package:mostro_mobile/shared/providers/notifications_history_repository_provider.dart';
 import 'package:mostro_mobile/data/enums.dart';
-
-class NotificationTemporaryState {
-  final Action? action;
-  final Map<String, dynamic> values;
-  final bool show;
-
-  NotificationTemporaryState({
-    this.action,
-    this.values = const {},
-    this.show = false,
-  });
-}
-
-class NotificationsState {
-  final AsyncValue<List<NotificationModel>> historyNotifications;
-  final NotificationTemporaryState temporaryNotification;
-
-  NotificationsState({
-    required this.historyNotifications,
-    required this.temporaryNotification,
-  });
-
-  NotificationsState copyWith({
-    AsyncValue<List<NotificationModel>>? historyNotifications,
-    NotificationTemporaryState? temporaryNotification,
-  }) {
-    return NotificationsState(
-      historyNotifications: historyNotifications ?? this.historyNotifications,
-      temporaryNotification: temporaryNotification ?? this.temporaryNotification,
-    );
-  }
-}
 
 class NotificationsNotifier extends StateNotifier<NotificationsState> {
   final Ref ref;
   late final NotificationsRepository _repository;
   
-  NotificationsNotifier(this.ref) : super(NotificationsState(
-    historyNotifications: const AsyncValue.loading(),
+  NotificationsNotifier(this.ref) : super(const NotificationsState(
+    historyNotifications: AsyncValue.loading(),
     temporaryNotification: NotificationTemporaryState(),
   )) {
     _repository = ref.read(notificationsRepositoryProvider);
@@ -119,7 +90,7 @@ class NotificationsNotifier extends StateNotifier<NotificationsState> {
 
   void clearTemporary() {
     state = state.copyWith(
-      temporaryNotification: NotificationTemporaryState(),
+      temporaryNotification: const NotificationTemporaryState(),
     );
   }
 
@@ -128,8 +99,8 @@ class NotificationsNotifier extends StateNotifier<NotificationsState> {
       id: DateTime.now().millisecondsSinceEpoch.toString(),
       type: NotificationModel.getNotificationTypeFromAction(action),
       action: action,
-      title: _getNotificationTitleKey(action),
-      message: _getNotificationMessageKey(action),
+      title: NotificationMessageMapper.getTitleKey(action),
+      message: NotificationMessageMapper.getMessageKey(action),
       timestamp: DateTime.now(),
       orderId: orderId,
       data: values,
@@ -142,139 +113,5 @@ class NotificationsNotifier extends StateNotifier<NotificationsState> {
     await addToHistory(action, values: values, orderId: orderId);
   }
 
-  String _getNotificationTitleKey(Action action) {
-    switch (action) {
-      case Action.newOrder:
-        return 'notification_new_order_title';
-      case Action.takeBuy:
-      case Action.takeSell:
-        return 'notification_order_taken_title';
-      case Action.payInvoice:
-        return 'notification_payment_required_title';
-      case Action.fiatSent:
-        return 'notification_fiat_sent_title';
-      case Action.fiatSentOk:
-        return 'notification_fiat_sent_ok_title';
-      case Action.release:
-        return 'notification_release_title';
-      case Action.released:
-        return 'notification_bitcoin_released_title';
-      case Action.buyerInvoiceAccepted:
-        return 'notification_buyer_invoice_accepted_title';
-      case Action.purchaseCompleted:
-        return 'notification_purchase_completed_title';
-      case Action.holdInvoicePaymentAccepted:
-        return 'notification_hold_invoice_payment_accepted_title';
-      case Action.holdInvoicePaymentSettled:
-        return 'notification_hold_invoice_payment_settled_title';
-      case Action.holdInvoicePaymentCanceled:
-        return 'notification_hold_invoice_payment_canceled_title';
-      case Action.waitingSellerToPay:
-        return 'notification_waiting_seller_to_pay_title';
-      case Action.waitingBuyerInvoice:
-        return 'notification_waiting_buyer_invoice_title';
-      case Action.addInvoice:
-        return 'notification_add_invoice_title';
-      case Action.buyerTookOrder:
-        return 'notification_buyer_took_order_title';
-      case Action.rate:
-      case Action.rateUser:
-        return 'notification_rate_title';
-      case Action.rateReceived:
-        return 'notification_rate_received_title';
-      case Action.dispute:
-        return 'notification_dispute_started_title';
-      case Action.disputeInitiatedByYou:
-        return 'notification_dispute_initiated_by_you_title';
-      case Action.disputeInitiatedByPeer:
-        return 'notification_dispute_initiated_by_peer_title';
-      case Action.paymentFailed:
-        return 'notification_payment_failed_title';
-      case Action.invoiceUpdated:
-        return 'notification_invoice_updated_title';
-      case Action.cantDo:
-        return 'notification_cant_do_title';
-      case Action.canceled:
-        return 'notification_order_canceled_title';
-      case Action.cooperativeCancelInitiatedByYou:
-        return 'notification_cooperative_cancel_initiated_by_you_title';
-      case Action.cooperativeCancelInitiatedByPeer:
-        return 'notification_cooperative_cancel_initiated_by_peer_title';
-      case Action.cooperativeCancelAccepted:
-        return 'notification_cooperative_cancel_accepted_title';
-      case Action.sendDm:
-        return 'notification_new_message_title';
-      default:
-        return 'notification_order_update_title';
-    }
-  }
-
-  String _getNotificationMessageKey(Action action) {
-    switch (action) {
-      case Action.newOrder:
-        return 'notification_new_order_message';
-      case Action.takeBuy:
-        return 'notification_sell_order_taken_message';
-      case Action.takeSell:
-        return 'notification_buy_order_taken_message';
-      case Action.payInvoice:
-        return 'notification_payment_required_message';
-      case Action.fiatSent:
-        return 'notification_fiat_sent_message';
-      case Action.fiatSentOk:
-        return 'notification_fiat_sent_ok_message';
-      case Action.release:
-        return 'notification_release_message';
-      case Action.released:
-        return 'notification_bitcoin_released_message';
-      case Action.buyerInvoiceAccepted:
-        return 'notification_buyer_invoice_accepted_message';
-      case Action.purchaseCompleted:
-        return 'notification_purchase_completed_message';
-      case Action.holdInvoicePaymentAccepted:
-        return 'notification_hold_invoice_payment_accepted_message';
-      case Action.holdInvoicePaymentSettled:
-        return 'notification_hold_invoice_payment_settled_message';
-      case Action.holdInvoicePaymentCanceled:
-        return 'notification_hold_invoice_payment_canceled_message';
-      case Action.waitingSellerToPay:
-        return 'notification_waiting_seller_to_pay_message';
-      case Action.waitingBuyerInvoice:
-        return 'notification_waiting_buyer_invoice_message';
-      case Action.addInvoice:
-        return 'notification_add_invoice_message';
-      case Action.buyerTookOrder:
-        return 'notification_buyer_took_order_message';
-      case Action.rate:
-      case Action.rateUser:
-        return 'notification_rate_message';
-      case Action.rateReceived:
-        return 'notification_rate_received_message';
-      case Action.dispute:
-        return 'notification_dispute_started_message';
-      case Action.disputeInitiatedByYou:
-        return 'notification_dispute_initiated_by_you_message';
-      case Action.disputeInitiatedByPeer:
-        return 'notification_dispute_initiated_by_peer_message';
-      case Action.paymentFailed:
-        return 'notification_payment_failed_message';
-      case Action.invoiceUpdated:
-        return 'notification_invoice_updated_message';
-      case Action.cantDo:
-        return 'notification_cant_do_message';
-      case Action.canceled:
-        return 'notification_order_canceled_message';
-      case Action.cooperativeCancelInitiatedByYou:
-        return 'notification_cooperative_cancel_initiated_by_you_message';
-      case Action.cooperativeCancelInitiatedByPeer:
-        return 'notification_cooperative_cancel_initiated_by_peer_message';
-      case Action.cooperativeCancelAccepted:
-        return 'notification_cooperative_cancel_accepted_message';
-      case Action.sendDm:
-        return 'notification_new_message_message';
-      default:
-        return 'notification_order_update_message';
-    }
-  }
 
 }
